+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* BUGFIX: draw_legend fails when classes are provided (@lwasser #253)
 * Remove earthpy.utils.fix_paths as it is not used in the package (@lwasser #259)
 * Adding tests for hillshade and improved docs (@jpalomino #260)
 * Closing plots in tests (@lwasser #257)